TIP-842: Move Attribute type to Akeneo\Pim\Structure
