+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/attribute_types.yml
@@ -1,16 +1,16 @@
 parameters:
-    pim_catalog.attribute_type.identifier.class:       Pim\Bundle\CatalogBundle\AttributeType\IdentifierType
-    pim_catalog.attribute_type.text.class:             Pim\Bundle\CatalogBundle\AttributeType\TextType
-    pim_catalog.attribute_type.textarea.class:         Pim\Bundle\CatalogBundle\AttributeType\TextAreaType
-    pim_catalog.attribute_type.multiselect.class:      Pim\Bundle\CatalogBundle\AttributeType\OptionMultiSelectType
-    pim_catalog.attribute_type.simpleselect.class:     Pim\Bundle\CatalogBundle\AttributeType\OptionSimpleSelectType
-    pim_catalog.attribute_type.price_collection.class: Pim\Bundle\CatalogBundle\AttributeType\PriceCollectionType
-    pim_catalog.attribute_type.number.class:           Pim\Bundle\CatalogBundle\AttributeType\NumberType
-    pim_catalog.attribute_type.boolean.class:          Pim\Bundle\CatalogBundle\AttributeType\BooleanType
-    pim_catalog.attribute_type.date.class:             Pim\Bundle\CatalogBundle\AttributeType\DateType
-    pim_catalog.attribute_type.file.class:             Pim\Bundle\CatalogBundle\AttributeType\FileType
-    pim_catalog.attribute_type.image.class:            Pim\Bundle\CatalogBundle\AttributeType\ImageType
-    pim_catalog.attribute_type.metric.class:           Pim\Bundle\CatalogBundle\AttributeType\MetricType
+    pim_catalog.attribute_type.identifier.class:       Akeneo\Pim\Structure\Component\AttributeType\IdentifierType
+    pim_catalog.attribute_type.text.class:             Akeneo\Pim\Structure\Component\AttributeType\TextType
+    pim_catalog.attribute_type.textarea.class:         Akeneo\Pim\Structure\Component\AttributeType\TextAreaType
+    pim_catalog.attribute_type.multiselect.class:      Akeneo\Pim\Structure\Component\AttributeType\OptionMultiSelectType
+    pim_catalog.attribute_type.simpleselect.class:     Akeneo\Pim\Structure\Component\AttributeType\OptionSimpleSelectType
+    pim_catalog.attribute_type.price_collection.class: Akeneo\Pim\Structure\Component\AttributeType\PriceCollectionType
+    pim_catalog.attribute_type.number.class:           Akeneo\Pim\Structure\Component\AttributeType\NumberType
+    pim_catalog.attribute_type.boolean.class:          Akeneo\Pim\Structure\Component\AttributeType\BooleanType
+    pim_catalog.attribute_type.date.class:             Akeneo\Pim\Structure\Component\AttributeType\DateType
+    pim_catalog.attribute_type.file.class:             Akeneo\Pim\Structure\Component\AttributeType\FileType
+    pim_catalog.attribute_type.image.class:            Akeneo\Pim\Structure\Component\AttributeType\ImageType
+    pim_catalog.attribute_type.metric.class:           Akeneo\Pim\Structure\Component\AttributeType\MetricType
 
 services:
     pim_catalog.registry.attribute_type:
