+++ src/Akeneo/Pim/Structure/Component/AttributeType/AbstractAttributeType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\AttributeType;
+namespace Akeneo\Pim\Structure\Component\AttributeType;
 
 use Pim\Component\Catalog\AttributeTypeInterface;
 use Pim\Component\Catalog\AttributeTypes;
