+++ src/Akeneo/Pim/Structure/Component/AttributeType/BooleanType.php
@@ -1,31 +1,23 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\AttributeType;
+namespace Akeneo\Pim\Structure\Component\AttributeType;
 
 use Pim\Component\Catalog\AttributeTypes;
 
 /**
- * Identifier attribute type
+ * Boolean attribute type
  *
  * @author    Filips Alpe <filips@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class IdentifierType extends AbstractAttributeType
+class BooleanType extends AbstractAttributeType
 {
     /**
      * {@inheritdoc}
      */
     public function getName()
     {
-        return AttributeTypes::IDENTIFIER;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isUnique()
-    {
-        return true;
+        return AttributeTypes::BOOLEAN;
     }
 }
