+++ src/Akeneo/Pim/Structure/Component/AttributeType/IdentifierType.php
@@ -1,23 +1,31 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\AttributeType;
+namespace Akeneo\Pim\Structure\Component\AttributeType;
 
 use Pim\Component\Catalog\AttributeTypes;
 
 /**
- * Boolean attribute type
+ * Identifier attribute type
  *
  * @author    Filips Alpe <filips@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class BooleanType extends AbstractAttributeType
+class IdentifierType extends AbstractAttributeType
 {
     /**
      * {@inheritdoc}
      */
     public function getName()
     {
-        return AttributeTypes::BOOLEAN;
+        return AttributeTypes::IDENTIFIER;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isUnique()
+    {
+        return true;
     }
 }
