+++ src/Akeneo/Pim/Structure/Component/AttributeType/OptionMultiSelectType.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\AttributeType;
+namespace Akeneo\Pim\Structure\Component\AttributeType;
 
 use Pim\Component\Catalog\AttributeTypes;
 
 /**
- * Metric attribute type
+ * Multi options (select) attribute type
  *
- * @author    Filips Alpe <filips@akeneo.com>
+ * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MetricType extends AbstractAttributeType
+class OptionMultiSelectType extends AbstractAttributeType
 {
     /**
      * {@inheritdoc}
      */
     public function getName()
     {
-        return AttributeTypes::METRIC;
+        return AttributeTypes::OPTION_MULTI_SELECT;
     }
 }
