+++ src/Akeneo/Pim/Structure/Component/AttributeType/OptionSimpleSelectType.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\AttributeType;
+namespace Akeneo\Pim\Structure\Component\AttributeType;
 
 use Pim\Component\Catalog\AttributeTypes;
 
 /**
- * Multi options (select) attribute type
+ * Simple options (select) attribute type
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class OptionMultiSelectType extends AbstractAttributeType
+class OptionSimpleSelectType extends AbstractAttributeType
 {
     /**
      * {@inheritdoc}
      */
     public function getName()
     {
-        return AttributeTypes::OPTION_MULTI_SELECT;
+        return AttributeTypes::OPTION_SIMPLE_SELECT;
     }
 }
