+++ src/Akeneo/Pim/Structure/Component/AttributeType/PriceCollectionType.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\AttributeType;
+namespace Akeneo\Pim\Structure\Component\AttributeType;
 
 use Pim\Component\Catalog\AttributeTypes;
 
 /**
- * Number attribute type
+ * Price attribute type
  *
- * @author    Filips Alpe <filips@akeneo.com>
+ * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class NumberType extends AbstractAttributeType
+class PriceCollectionType extends AbstractAttributeType
 {
     /**
      * {@inheritdoc}
      */
     public function getName()
     {
-        return AttributeTypes::NUMBER;
+        return AttributeTypes::PRICE_COLLECTION;
     }
 }
