+++ src/Akeneo/Pim/Structure/Component/AttributeType/TextType.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\AttributeType;
+namespace Akeneo\Pim\Structure\Component\AttributeType;
 
 use Pim\Component\Catalog\AttributeTypes;
 
 /**
- * Date attribute type
+ * Text attribute type
  *
  * @author    Filips Alpe <filips@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class DateType extends AbstractAttributeType
+class TextType extends AbstractAttributeType
 {
     /**
      * {@inheritdoc}
      */
     public function getName()
     {
-        return AttributeTypes::DATE;
+        return AttributeTypes::TEXT;
     }
 }
