+++ src/Akeneo/Pim/Structure/Component/spec/Factory/AttributeFactorySpec.php
@@ -3,7 +3,7 @@
 namespace spec\Akeneo\Pim\Structure\Component\Factory;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\AttributeType\AbstractAttributeType;
+use Akeneo\Pim\Structure\Component\AttributeType\AbstractAttributeType;
 use Akeneo\Pim\Structure\Component\AttributeTypeRegistry;
 
 class AttributeFactorySpec extends ObjectBehavior
